build(lint): bump ruff pin to 0.16.0 + apply its safe fixes
ruff 0.16 expanded its default rule set (~850 findings tree-wide, none
behaviour regressions).  CI pinned 0.15.12 to dodge them; this bumps the
pin and cleans the tree.  This commit is the mechanical half:

  * ci.yml: ruff 0.15.12 -> 0.16.0
  * pyproject: fix target-version py311 -> py39 to match requires-python.
    The old py311 made ruff rewrite annotations to 3.10/3.11-only syntax
    (PEP 604 `X | None`, `datetime.UTC`) that crashes at import on 3.9,
    because Pydantic models and Typer commands resolve annotations via
    get_type_hints() even under `from __future__ import annotations`.
  * pyproject: ignore UP006/UP007/UP045 while we still support 3.9
    (keep typing.Optional/Union/List), with a note to drop the block and
    re-run --fix when the 3.9 floor is raised.
  * apply `ruff check . --fix` (safe fixes only): import sorting, unused
    noqa, quoted-annotation removal, regex-flag aliases, etc.

The ~196 findings needing per-site judgment (BLE001, B008, S110, DTZ*, ...)
are handled in the follow-up commit.

Verified on Python 3.9.13: pytest tests/unit green (775 passed, 9 skipped),
mypy bioflow --ignore-missing-imports clean.

+++ tests/unit/test_doctor.py
@@ -12,7 +12,6 @@ from bioflow.core.doctor import (
     summarize,
 )
 
-
 # ---------------------------------------------------------------------------
 # Pure check functions
 # ---------------------------------------------------------------------------
@@ -107,7 +106,7 @@ class TestIndividualChecks:
         import psutil
 
         class _Mem:
-            total = int(2 * 1024**3)  # 2 GB
+            total = 2 * 1024**3  # 2 GB
 
         monkeypatch.setattr(psutil, "virtual_memory", lambda: _Mem())
         r = doctor._check_ram()
@@ -117,7 +116,7 @@ class TestIndividualChecks:
         import psutil
 
         class _Mem:
-            total = int(6 * 1024**3)  # 6 GB
+            total = 6 * 1024**3  # 6 GB
 
         monkeypatch.setattr(psutil, "virtual_memory", lambda: _Mem())
         r = doctor._check_ram()
